Update names for the observe package
