+++ pkg/kstatus/polling/event/eventtype_string.go
